feat 초기설정-1
