+++ src/main/java/com/example/letmovie/LetMovieApplication.java
@@ -1,0 +1,13 @@
+package com.example.letmovie;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LetMovieApplication  {
+
+    public static void main(String[] args) {
+        SpringApplication.run(LetMovieApplication .class, args);
+    }
+
+}
