+++ src/main/java/com/example/letmovie/domain/board/entity/Like.java
@@ -1,0 +1,4 @@
+package com.example.letmovie.domain.board.entity;
+
+public class Like {
+}
