+++ src/main/java/com/example/letmovie/domain/board/entity/Post.java
@@ -1,0 +1,4 @@
+package com.example.letmovie.domain.board.entity;
+
+public class Post {
+}
