+++ src/main/java/com/example/letmovie/domain/member/entity/Grade.java
@@ -1,0 +1,6 @@
+package com.example.letmovie.domain.member.entity;
+
+public enum Grade {
+    GENERAL, VIP
+
+}
