+++ src/main/java/com/example/letmovie/domain/member/entity/Member.java
@@ -1,0 +1,32 @@
+package com.example.letmovie.domain.member.entity;
+
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "member_id")
+    private Long id;
+
+    @Column(nullable = false)
+    private String nickname;
+    @Column(nullable = false)
+    private String email;
+    @Column(nullable = false)
+    private String password;
+    @Column(nullable = false)
+    private String zipCode;
+    @Column(nullable = false)
+    private boolean isAdmin;
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private Grade grade;
+}
