+++ src/main/java/com/example/letmovie/domain/movie/entity/ShowTime.java
@@ -1,0 +1,4 @@
+package com.example.letmovie.domain.movie.entity;
+
+public class ShowTime {
+}
