+++ src/main/java/com/example/letmovie/domain/reservation/entity/ReservationStatus.java
@@ -1,0 +1,9 @@
+package com.example.letmovie.domain.reservation.entity;
+
+public enum ReservationStatus {
+
+    PENDING,       // 예약 중
+    CANCELLED,     // 취소
+    COMPLETED,     // 사용 완료
+
+    }
