+++ src/main/java/com/example/letmovie/domain/reservation/entity/Seat.java
@@ -1,0 +1,32 @@
+package com.example.letmovie.domain.reservation.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Seat {
+
+    @Id
+    @Column(name ="seat_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "screen_id")
+    private Screen screen;
+
+    @Enumerated(EnumType.STRING) //enum
+    private SeatType seatType;
+
+    private int seatLow;
+    private int seatCol;
+
+    private boolean isAble;
+    private int price;
+
+
+}
