+++ src/main/java/com/example/letmovie/domain/reservation/entity/SeatType.java
@@ -1,0 +1,8 @@
+package com.example.letmovie.domain.reservation.entity;
+
+public enum SeatType {
+    VIP,         // VIP 좌석
+    PREMIUM,     // 프리미엄 좌석
+    REGULAR,     // 일반 좌석
+    ECONOMY      // 저가 좌석
+}
