--- src/main/java/com/example/nbe232team2/domain/board/entity/Comment.java
@@ -1,4 +1,0 @@
-package com.example.nbe232team2.domain.board.entity;
-
-public class Comment {
-}
