--- src/main/java/com/example/nbe232team2/domain/movie/entity/Movie.java
@@ -1,4 +1,0 @@
-package com.example.nbe232team2.domain.movie.entity;
-
-public class Movie {
-}
