--- src/main/java/com/example/nbe232team2/domain/movie/entity/Theater.java
@@ -1,4 +1,0 @@
-package com.example.nbe232team2.domain.movie.entity;
-
-public class Theater {
-}
