--- src/main/java/com/example/nbe232team2/domain/reservation/entity/Reservation.java
@@ -1,6 +1,0 @@
-package com.example.nbe232team2.domain.reservation.entity;
-
-public class Reservation {
-
-
-}
