--- src/main/java/com/example/nbe232team2/domain/reservation/entity/ReservationSeat.java
@@ -1,4 +1,0 @@
-package com.example.nbe232team2.domain.reservation.entity;
-
-public class ReservationSeat {
-}
