--- src/main/java/com/example/nbe232team2/domain/reservation/entity/Screen.java
@@ -1,6 +1,0 @@
-package com.example.nbe232team2.domain.reservation.entity;
-
-public class Screen {
-
-
-}
