+++ src/main/resources/application.properties
@@ -1,9 +1,18 @@
 spring.application.name=NBE2-3-2-team2
 
 
-spring.jpa.hibernate.ddl-auto=create
-spring.datasource.driverClassName=org.mariadb.jdbc.Driver
-spring.datasource.url="jdbc:mariadb://localhost:3306/letmovie"
-spring.datasource.username="root" 
-spring.datasource.password=""
+#spring.jpa.hibernate.ddl-auto=create
+#spring.datasource.driverClassName=org.mariadb.jdbc.Driver
+#spring.datasource.url="jdbc:mariadb://localhost:3306/letmovie"
+#spring.datasource.username="root"
+#spring.datasource.password=""
+#
+spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
+spring.datasource.url=jdbc:mariadb://localhost:3306/sample
+spring.datasource.username=root
+spring.datasource.password=!123456
 
+spring.jpa.database-platform=org.hibernate.dialect.MariaDBDialect
+
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.show-sql = true
