+++ src/test/java/com/example/letmovie/Nbe232Team2ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.letmovie;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Nbe232Team2ApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
