--- src/test/java/com/example/nbe232team2/Nbe232Team2ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.nbe232team2;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Nbe232Team2ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
